Use the GeminiColor alpha selector wherever applicable
Now that GeminiConfig is fixed for GeminiColor and the hasAlpha flag for the widget switch away from separate opacity sliders wherever it is applicable. Basically everywhere except at the MBTelegraphAssist because you probably want to change that independent of the PP count colors.
This also fixed the opacity not actually working for when PP colors were used for the MBTelegraphAssist colors.

Removed the remaining unnecesary getters and setters ( should have used regexp in the first place to clean it up )